centralize mobile logic to hook
